:art: 重构 main
