+++ test/widget_test.dart
@@ -25,6 +25,6 @@ import 'package:pixez/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    await tester.pumpWidget(MyApp());
+    await tester.pumpWidget(App());
   });
 }
